requirements.txt
Flask
Flask-SQLAlchemy
Flask-Admin
google-generativeai
APScheduler
python-dotenv